\encoding{UTF-8}
\name{aptitude}
\Rdversion{1.1}
\alias{aptitude}
\docType{data}
\title{
Sample dataset: aptitude
}
\description{
Data of two samples of testees who completed an aptitude test consisting of general knowledge questions,
      logic tasks, and two measures of intelligence.
}
\usage{data("aptitude")}
\format{
  A list that contains two data.frames holding 291 and 334 observations on the following 4 variables.
  \describe{
    \item{\code{knowledge}}{Score achieved on the general knowledge questions (numeric vector)}
    \item{\code{logic}}{Score achieved on the logic tasks (numeric vector)}
    \item{\code{intelligence.a}}{Intelligence measure A (numeric vector)}
    \item{\code{intelligence.b}}{Intelligence measure B (numeric vector)}
  }
}
\examples{
data("aptitude")
}
\keyword{datasets}
