% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocor-package.R
\docType{package}
\name{cocor-package}
\alias{cocor-package}
\title{\packageTitle{cocor}}
\description{
\packageDescription{cocor}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab cocor\cr
Type: \tab Package\cr
Version: \tab 1.1-4\cr
Date: \tab 2022-06-28\cr
Depends: \tab methods\cr
Enhances: \tab rkward\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://comparingcorrelations.org\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{http://comparingcorrelations.org}
}

}
\author{
\packageAuthor{cocor}

Maintainer: \packageMaintainer{cocor}
}
