\docType{package}
\name{cocor-package}
\alias{cocor}
\alias{cocor-package}
\title{The cocor Package}
\description{
  Comparing correlations.
}
\details{
  \tabular{ll}{ Package: \tab cocor\cr Type: \tab
  Package\cr Version: \tab 0.01-2\cr Date: \tab
  2013-02-20\cr Depends: \tab methods\cr Enhances: \tab
  rkward\cr Encoding: \tab UTF-8\cr License: \tab GPL (>=
  3)\cr LazyLoad: \tab yes\cr URL: \tab
  http://r.birkdiedenhofen.de/pckg/cocor/\cr }

  This package provides functions to compare two
  correlations based on either dependent or independent
  groups.
}
\author{
  Birk Diedenhofen \email{mail@birkdiedenhofen.de}
}
\keyword{package}

